Testrun to target: remote33_dienstag1

+++ src/de/lmu/dbs/ciaa/classifier/features/MCFeature1.java
@@ -89,6 +89,6 @@ public class MCFeature1 extends MCFeature {
 
 	@Override
 	public float getMaxValue() {
-		return Byte.MAX_VALUE-1;
+		return (float)(Byte.MAX_VALUE-1);
 	}
 }
